+++ rust/src/producers/builder.rs
@@ -1,9 +1,10 @@
 use std::mem::take;
+use std::collections::HashMap;
 
-pub use super::build_gates::BuildGate;
+pub use super::build_gates::{BuildGate, BuildComplexGate};
 use super::build_gates::NO_OUTPUT;
 use crate::producers::sink::MemorySink;
-use crate::structs::value::Value;
+use crate::structs::{value::Value, function::Function, wire::WireList, wire::WireListElement};
 use crate::{Gate, Header, Instance, Relation, Sink, WireId, Witness};
 use std::cell::{RefCell, Cell};
 use crate::structs::relation::{ARITH, FUNCTION, FOR, SWITCH};
@@ -12,6 +13,10 @@ pub trait GateBuilderT {
     /// Allocates a new wire id for the output and creates a new gate,
     /// Returns the newly allocated WireId.
     fn create_gate(&self, gate: BuildGate) -> WireId;
+
+    /// Allocates some new wire ids for the output and creates a new gate,
+    /// Returns the newly allocated WireIds.
+    fn create_complex_gate(&self, gate: BuildComplexGate) -> Result<WireList, String>;
 }
 
 /// MessageBuilder builds messages by buffering sequences of gates and witness/instance values.
@@ -70,7 +75,14 @@ impl<S: Sink> MessageBuilder<S> {
 
     fn push_gate(&mut self, gate: Gate) {
         self.relation.gates.push(gate);
-        if self.relation.gates.len() == self.max_len {
+        if self.relation.gates.len() + self.relation.functions.len() == self.max_len {
+            self.flush_relation();
+        }
+    }
+
+    fn push_function(&mut self, function: Function) {
+        self.relation.functions.push(function);
+        if self.relation.gates.len() + self.relation.functions.len() == self.max_len {
             self.flush_relation();
         }
     }
@@ -88,6 +100,7 @@ impl<S: Sink> MessageBuilder<S> {
     fn flush_relation(&mut self) {
         self.sink.push_relation_message(&self.relation).unwrap();
         self.relation.gates.clear();
+        self.relation.functions.clear();
     }
 
     fn finish(mut self) -> S {
@@ -97,7 +110,7 @@ impl<S: Sink> MessageBuilder<S> {
         if !self.witness.short_witness.is_empty() {
             self.flush_witness();
         }
-        if !self.relation.gates.is_empty() {
+        if !self.relation.gates.is_empty() || !self.relation.functions.is_empty() {
             self.flush_relation();
         }
         self.sink
@@ -112,7 +125,7 @@ impl<S: Sink> MessageBuilder<S> {
 /// use zki_sieve::producers::sink::MemorySink;
 /// use zki_sieve::Header;
 ///
-/// let mut b = GateBuilder::new(MemorySink::default(), Header::default());
+/// let b = GateBuilder::new(MemorySink::default(), Header::default());
 ///
 /// let my_id = b.create_gate(Constant(vec![0]));
 /// b.create_gate(AssertZero(my_id));
@@ -120,6 +133,7 @@ impl<S: Sink> MessageBuilder<S> {
 pub struct GateBuilder<S: Sink> {
     msg_build: RefCell<MessageBuilder<S>>,
 
+    known_functions: RefCell<HashMap<String, usize>>,
     free_id: Cell<WireId>,
 }
 
@@ -145,6 +159,17 @@ impl<S: Sink> GateBuilderT for GateBuilder<S> {
 
         out_id
     }
+
+    fn create_complex_gate(&self, gate: BuildComplexGate) -> Result<WireList, String> {
+        let output_count = match gate {
+            BuildComplexGate::Call(ref name, _) => self.known_function_output_count(name)?,
+        };
+        let output_wires = self.multiple_alloc(output_count);
+
+        self.msg_build.borrow_mut().push_gate(gate.with_output(output_wires.clone()));
+
+        Ok(output_wires)
+    }
 }
 
 impl<S: Sink> GateBuilder<S> {
@@ -156,6 +181,7 @@ impl<S: Sink> GateBuilder<S> {
     pub fn new_with_functionalities(sink: S, header: Header, gateset: u16, features: u16) -> Self {
         GateBuilder {
             msg_build: RefCell::new(MessageBuilder::new(sink, header, gateset, features)),
+            known_functions: RefCell::new(HashMap::new()),
             free_id: Cell::new(0),
         }
     }
@@ -167,12 +193,38 @@ impl<S: Sink> GateBuilder<S> {
         id
     }
 
-    pub(crate) fn push_witness_value(&self, val: Value) {
+    /// alloc allocates n wire IDs.
+    fn multiple_alloc(&self, n: usize) -> WireList {
+        let id = self.free_id.take();
+        let next: u64 = id + n as u64;
+        self.free_id.set(next);
+        vec![WireListElement::WireRange(id, next-1)]
+    }
+
+    /// known_function_output_count returns the output_count of the function with name `name`.
+    fn known_function_output_count(&self, name: &String) -> Result<usize, String> {
+        match self.known_functions.borrow().get(name) {
+            None => Err(format!("Function {} does not exist !", name)),
+            Some(v) => Ok(*v),
+        }
+    }
+
+    pub fn push_witness_value(&self, val: Value) {
         self.msg_build.borrow_mut().push_witness_value(val);
     }
 
-    pub(crate) fn push_instance_value(&self, val: Value) {
+    pub fn push_instance_value(&self, val: Value) {
         self.msg_build.borrow_mut().push_instance_value(val);
+    }
+
+    pub fn push_function(&self, function: Function) -> Result<(), String> {
+        if self.known_functions.borrow().contains_key(&function.name) {
+            return Err(format!("Function {} already exists !", function.name));
+        } else {
+            self.known_functions.borrow_mut().insert(function.name.clone(), function.output_count);
+        }
+        self.msg_build.borrow_mut().push_function(function);
+        Ok(())
     }
 
     pub fn finish(self) -> S {
@@ -182,4 +234,121 @@ impl<S: Sink> GateBuilder<S> {
 
 pub fn new_example_builder() -> GateBuilder<MemorySink> {
     GateBuilder::new(MemorySink::default(), Header::default())
+}
+
+#[test]
+fn test_builder_with_function() {
+    use crate::producers::builder::{GateBuilderT, GateBuilder, BuildGate::*, BuildComplexGate::*};
+    use crate::producers::{sink::MemorySink, examples};
+    use crate::structs::{gates::Gate, function::Function};
+    use crate::structs::wire::{WireListElement::*, expand_wirelist};
+    use crate::consumers::source::Source;
+    use crate::consumers::evaluator::{Evaluator, PlaintextBackend};
+
+    let b = GateBuilder::new(MemorySink::default(), examples::example_header());
+
+    let custom_sub = Function::new(
+        "custom_sub".to_string(),
+        2,
+        4,
+        0,
+        0,
+        vec![
+            Gate::MulConstant(6, 4, vec![100]), // *(-1)
+            Gate::MulConstant(7, 5, vec![100]), // *(-1)
+            Gate::Add(0, 2, 6),
+            Gate::Add(1, 3, 7)]
+     );
+    b.push_function(custom_sub).unwrap();
+
+    // Try to push two functions with the same name
+    // It should return an error
+    let custom_function= Function::new(
+        "custom_sub".to_string(),
+        0,
+        0,
+        0,
+        0,
+        vec![]
+    );
+    assert!(b.push_function(custom_function).is_err());
+
+    let id_0 = b.create_gate(Constant(vec![40]));
+    let id_1 = b.create_gate(Constant(vec![30]));
+    let id_2 = b.create_gate(Constant(vec![10]));
+    let id_3 = b.create_gate(Constant(vec![5]));
+
+    let out = b.create_complex_gate(Call("custom_sub".to_string(), vec![Wire(id_0), Wire(id_1), Wire(id_2), Wire(id_3)])).unwrap();
+    let out = expand_wirelist(&out);
+    assert_eq!(out.len(), 2);
+
+    let witness_0 = b.create_gate(Witness(Some(vec![30])));
+    let witness_1 = b.create_gate(Witness(Some(vec![25])));
+
+    let neg_witness_0 = b.create_gate(MulConstant(witness_0, vec![100])); // *(-1)
+    let neg_witness_1 = b.create_gate(MulConstant(witness_1, vec![100])); // *(-1)
+
+    let res_0 = b.create_gate(Add(out[0], neg_witness_0));
+    let res_1 = b.create_gate(Add(out[1], neg_witness_1));
+
+    b.create_gate(AssertZero(res_0));
+    b.create_gate(AssertZero(res_1));
+
+    // Try to call an unknown function
+    // It should return an error
+    assert!(b.create_complex_gate(Call("unknown_function".to_string(), vec![Wire(id_0)])).is_err());
+
+    let sink = b.finish();
+
+    let mut zkbackend = PlaintextBackend::default();
+    let source: Source = sink.into();
+    let evaluator = Evaluator::from_messages(source.iter_messages(), &mut zkbackend);
+    assert_eq!(evaluator.get_violations().len(), 0);
+}
+
+#[test]
+fn test_builder_with_witness_instance_function() {
+    use crate::producers::builder::{GateBuilderT, GateBuilder, BuildGate::*, BuildComplexGate::*};
+    use crate::producers::{sink::MemorySink, examples};
+    use crate::structs::{gates::Gate, function::Function};
+    use crate::structs::wire::{WireListElement::*, expand_wirelist};
+    use crate::consumers::source::Source;
+    use crate::consumers::evaluator::{Evaluator, PlaintextBackend};
+
+    let b = GateBuilder::new(MemorySink::default(), examples::example_header());
+
+    let sub_witness_instance = Function::new(
+        "sub_witness_instance".to_string(),
+        2,
+        2,
+        1,
+        1,
+        vec![
+            Gate::Instance(4),
+            Gate::Witness(5),
+            Gate::MulConstant(6, 4, vec![100]), // -instance
+            Gate::MulConstant(7, 5, vec![100]), // -witness
+            Gate::Add(0, 2, 6), // input[0] - instance
+            Gate::Add(1, 3, 7)] // input[1] - witness
+    );
+    b.push_function(sub_witness_instance).unwrap();
+
+    b.push_instance_value(vec![10]);
+    b.push_witness_value(vec![10]);
+
+    let id_0 = b.create_gate(Constant(vec![10]));
+    let id_1 = b.create_gate(Constant(vec![10]));
+
+    let out = b.create_complex_gate(Call("sub_witness_instance".to_string(), vec![Wire(id_0), Wire(id_1)])).unwrap();
+    let out = expand_wirelist(&out);
+    assert_eq!(out.len(), 2);
+
+    b.create_gate(AssertZero(out[0]));
+    b.create_gate(AssertZero(out[1]));
+    let sink = b.finish();
+
+    let mut zkbackend = PlaintextBackend::default();
+    let source: Source = sink.into();
+    let evaluator = Evaluator::from_messages(source.iter_messages(), &mut zkbackend);
+    assert_eq!(evaluator.get_violations().len(), 0);
 }