+++ Changelog.md
@@ -6,6 +6,7 @@ Schema (.fbs):
 Rust:
 - Boolean example (`zki_sieve bool-example [--incorrect]`)
 - Move to_r1cs converter from producers to consumers
+- Add function (sub-circuit) into the builder
 
 # v2.0.0, 2021-12